requirements.txt
absl-py==1.4.0
anyio==3.6.2
appdirs==1.4.4
async-generator==1.10
attrs==22.2.0
autopep8==2.0.1
beautifulsoup4==4.11.2
bidict==0.22.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
cmake==3.25.2
colorama==0.4.6
contourpy==1.0.7
cors==1.0.1
cssselect==1.2.0
cycler==0.11.0
data-scraper==1.1.4
deep-translator==1.10.1
fake-useragent==1.1.1
fastapi==0.92.0
fastapi-socketio==0.0.9
filelock==3.9.0
fonttools==4.38.0
future==0.18.3
gevent==22.10.2
google-trans-new==1.1.9
googletrans==4.0.0rc1
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httptools==0.5.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.0.0
jax==0.2.22
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.2
markdown2==2.4.8
MarkupSafe==2.1.2
matplotlib==3.7.1
nicegui==1.1.9
numpy==1.24.1
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parse==1.19.0
Pillow==9.4.0
plotly==5.13.1
pscript==0.7.7
py-translate==1.0.3
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.5
pyee==8.2.2
Pygments==2.14.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
pyrvsignal==2023.2.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.3.4
python-multipart==0.0.5
python-socketio==5.7.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-file==1.5.1
requests-html==0.10.0
rfc3986==1.5.0
scipy==1.10.0
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
starlette==0.25.0
tenacity==8.2.2
tldextract==3.4.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0

urllib3==1.26.14
uvicorn==0.20.0
vbuild==0.8.1
w3lib==2.1.1
watchfiles==0.18.1
websockets==10.4
wsproto==1.2.0
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
